third_party/blink/web_tests/fast/frames/sandboxed-iframe-attribute-parsing-03-expected.txt

CONSOLE ERROR: Error while parsing the 'sandbox' attribute: 'För', 'var', 'vers,', 'jag', 'gör,', 'Lovar', 'du', 'en', 'kyss', 'mig', 'giva;', 'Arket', 'fullt', 'borde', 'skriva,', 'Mindre', 'har', 'skrivit', 'för.', 'Men', 'man', 'måste', 'hålla', 'måtta,', 'Jag', 'med', 'kyssar', 'du.', 'Låt', 'räkna:', 'Där', 'är', 'sju!', 'En', 'därtill', 'det', 'gör', 'åtta.', 'Numro', 'åtta', 'fatal,', 'Greklands', 'sångmör', 'voro', 'nio,', 'Och', 'svensk', 'tio.', '—', 'Elva', 'apostelns', 'tal,', 'Ty', 'räknar', 'icke', 'Judas,', 'Honom,', 'som', 'i', 'vänners', 'lag', 'Kysste', 'falskt;', 'ej', 'jag,', 'Helst', 'när', 'vackra', 'läppar', 'bjudas.', 'Huru', 'står', 'min', 'räkning', 'här?', 'Aderton;', 'dock', 'något.', 'Nitton', 'rimmet', 'besvär,', 'Därföre', 'fyller', 'tjoget.', 'Strofen', 'full', 'In', 'hamnen', 'vill', 'styra,', 'till', 'godo', 'tag', 'Denna', 'gång', 'tjugofyra.', ''Kyssarna'', '('The', 'kisses'),', 'Esaias', 'Tegnér,', '1782-1846', 'int', 'main(void)', '{', 'return', '0;', '}' are invalid sandbox flags.
CONSOLE WARNING: An iframe which has both allow-scripts and allow-same-origin for its sandbox attribute can escape its sandboxing.
CONSOLE ERROR: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
Ridiculously long, invalid text (well, for these purposes at least) with non-ASCII characters surrounding attribute value

PASS
